Add team logos

+++ scripts/clone_production_db.sh
@@ -1,4 +1,4 @@
-rm latest.dump &&
+[ ! -e file ] || rm latest.dump &&
 heroku pg:backups:capture &&
 heroku pg:backups:download &&
 pg_restore --verbose --clean --no-acl --no-owner -h localhost -d web_development latest.dump
